fix: 요약 요청 페이지 수정사항 반영 - 요약 요청 페이지 이유 선택 토글박스 팝업 기능 추가 - src/components/scenario/pre/page.tsx 서버 컴포넌트로 리팩토링 - src/components/scenario/pre/ScenarioQuestionBox.tsx 버튼 배열 부분 렌더링하여 재배열
