+++ src/components/scenario/pre/ScenarioQuestionBox.tsx
@@ -23,67 +23,53 @@ const ScenarioQuestionBox = ({consultType, setConsultType, reason, setReason}: S
         </div>
 
         {/* 2. 상담 종류 Badge 컴포넌트 */}
-        <div className="mt-[12px] flex flex-col gap-[8px]">
-
-          {/* 윗줄 */}
-          <div className="flex justify-between">
-            {CONSULT_TYPES.slice(0, 3).map((item) => (
+        <div className="mt-[12px] px-[5px] max-width-[300px] mx-auto">
+          <div className="flex flex-wrap justify-between gap-y-[10px]">
+            {CONSULT_TYPES.map((item) => (
               <button
                 key={item.key}
                 type="button"
-                aria-pressed={consultType === item.key}
+                className="block"
                 onClick={() => {
                   setConsultType(item.key);
                   setReason(null);
                 }}
               >
-                <div className={ consultType === item.key ? 'rounded-full ring-2 ring-[#FBE783]' : '' } >
-                  <Badge type="primary">{item.label}</Badge>
+                <div className={ consultType === item.key ? 'rounded-[20px] ring-2 ring-[#FBE783]' : '' }>
+                  <div className="whitespace-nowrap">
+                    <Badge type="primary">{item.label}</Badge>
+                  </div>
                 </div>
               </button>
             ))}
+            
           </div>
-
-          {/* 아랫줄 */}
-          <div className="flex justify-between">
-            {CONSULT_TYPES.slice(3, 6).map((item) => (
-              <button
-                key={item.key}
-                type="button"
-                aria-pressed={consultType === item.key}
-                onClick={() => {
-                  setConsultType(item.key);
-                  setReason(null);
-                }}
-              >
-                <div className={ consultType === item.key ? 'rounded-full ring-2 ring-[#FBE783]' : '' } >
-                  <Badge type="primary">{item.label}</Badge>
-                </div>
-              </button>
-            ))}
-          </div>
-
         </div>
 
-        {/* 3. 이유 질문 */}
-        <div className="mt-[53px]">
-          <p className="heading3">변경하시는 이유가 무엇인가요?</p>
-        </div>
+        {consultType && (
+          <>
+            {/* 3. 이유 질문 */}
+            <div className="mt-[53px]">
+              <p className="heading3">변경하시는 이유가 무엇인가요?</p>
+            </div>
 
-        {/* 4. 이유 선택 Dropdown */}
-        <div className="mt-[12px]">
-          <Dropdown
-            value={reason}
-            onChange={setReason}
-            options={consultType ? REASONS_BY_TYPE[consultType] : []}
-            placeholder={
-              consultType
-                ? '이유를 선택하세요.'
-                : '상담 종류를 먼저 선택하세요.'
-            }
-            width="w-[300px]"
-          />
-        </div>
+            {/* 4. 이유 선택 Dropdown */}
+            <div className="mt-[12px]">
+              <Dropdown
+                value={reason}
+                onChange={setReason}
+                options={consultType ? REASONS_BY_TYPE[consultType] : []}
+                placeholder={
+                  consultType
+                    ? '이유를 선택하세요.'
+                    : '상담 종류를 먼저 선택하세요.'
+                }
+                width="w-[300px]"
+              />
+              </div>
+            </>
+        )}
+
 
       </div>
     </div>
